+++ packages/components-react/src/pages/index.tsx
@@ -69,11 +69,12 @@ import { TextareaWrapperPage } from './generated/TextareaWrapper';
 import { TextFieldWrapperPage } from './generated/TextFieldWrapper';
 import { TextListPage } from './generated/TextList';
 import { ToastPage } from './generated/Toast';
-import { ToastBasicDarkPage } from './generated/ToastBasicDark';
-import { ToastBasicLongTextPage } from './generated/ToastBasicLongText';
-import { ToastBasicStateNeutralPage } from './generated/ToastBasicStateNeutral';
+import { ToastBasicPage } from './generated/ToastBasic';
+import { ToastMultilinePage } from './generated/ToastMultiline';
 import { ToastOffsetPage } from './generated/ToastOffset';
 import { ToastPrefixedPage } from './generated/ToastPrefixed';
+import { ToastStateNeutralPage } from './generated/ToastStateNeutral';
+import { ToastStateSuccessPage } from './generated/ToastStateSuccess';
 import { TypographyCyrilPage } from './generated/TypographyCyril';
 import { TypographyFallbackPage } from './generated/TypographyFallback';
 import { TypographyGreekAndCopticPage } from './generated/TypographyGreekAndCoptic';
@@ -363,19 +364,14 @@ export const generatedRoutes: RouteType[] = [
     element: <ToastPage />,
   },
   {
-    name: 'Toast Basic Dark',
-    path: '/toast-basic-dark',
-    element: <ToastBasicDarkPage />,
+    name: 'Toast Basic',
+    path: '/toast-basic',
+    element: <ToastBasicPage />,
   },
   {
-    name: 'Toast Basic Long Text',
-    path: '/toast-basic-long-text',
-    element: <ToastBasicLongTextPage />,
-  },
-  {
-    name: 'Toast Basic State Neutral',
-    path: '/toast-basic-state-neutral',
-    element: <ToastBasicStateNeutralPage />,
+    name: 'Toast Multiline',
+    path: '/toast-multiline',
+    element: <ToastMultilinePage />,
   },
   {
     name: 'Toast Offset',
@@ -386,6 +382,16 @@ export const generatedRoutes: RouteType[] = [
     name: 'Toast Prefixed',
     path: '/toast-prefixed',
     element: <ToastPrefixedPage />,
+  },
+  {
+    name: 'Toast State Neutral',
+    path: '/toast-state-neutral',
+    element: <ToastStateNeutralPage />,
+  },
+  {
+    name: 'Toast State Success',
+    path: '/toast-state-success',
+    element: <ToastStateSuccessPage />,
   },
   {
     name: 'Typography Cyril',