+++ packages/components/src/components/flyout-navigation/flyout-navigation/flyout-navigation-utils.ts
@@ -30,11 +30,34 @@ export const validateActiveIdentifier = <T extends Class<any>>(
   items: HTMLPFlyoutNavigationItemElement[],
   activeIdentifier: string | undefined
 ): void => {
-  if (!(activeIdentifier === undefined || !!items.filter((item) => item.identifier === activeIdentifier).length)) {
-    consoleError(
-      `Invalid value '${activeIdentifier}' supplied to ${getTagNameWithoutPrefix(
-        instance.host as HTMLElement
-      )} for property 'activeIdentifier' because reference is not present.`
-    );
+  if (activeIdentifier !== undefined) {
+    const matchingItems = items.filter((item) => item.identifier === activeIdentifier);
+    if (matchingItems.length === 0) {
+      logInvalidIdentifierError(instance, activeIdentifier);
+    } else if (matchingItems.length > 1) {
+      logMultipleIdentifierError(instance, activeIdentifier, matchingItems);
+    }
   }
 };
+
+const logInvalidIdentifierError = <T extends Class<any>>(
+  instance: InstanceType<T>,
+  activeIdentifier: string | undefined
+) =>
+  consoleError(
+    `Invalid value '${activeIdentifier}' supplied to ${getTagNameWithoutPrefix(
+      instance.host as HTMLElement
+    )} for property 'activeIdentifier' because reference is not present.`
+  );
+
+const logMultipleIdentifierError = <T extends Class<any>>(
+  instance: InstanceType<T>,
+  activeIdentifier: string | undefined,
+  matchingItems: HTMLElement[]
+) =>
+  consoleError(
+    `Found multiple matching items for value '${activeIdentifier}' supplied to ${getTagNameWithoutPrefix(
+      instance.host as HTMLElement
+    )}:`,
+    ...matchingItems
+  );