+++ packages/components/src/components/flyout-navigation/flyout-navigation/flyout-navigation.tsx
@@ -1,4 +1,4 @@
-import { Component, Element, Event, type EventEmitter, h, type JSX, Prop, Watch } from '@stencil/core';
+import { Component, Element, Event, type EventEmitter, h, type JSX, Prop, Watch, State } from '@stencil/core';
 import {
   FLYOUT_NAVIGATION_ARIA_ATTRIBUTES,
   type FlyoutNavigationAriaAttribute,
@@ -42,7 +42,7 @@ export class FlyoutNavigation {
   @Prop() public open?: boolean = false;
 
   /** Defines which flyout-navigation-item to be visualized as opened. */
-  @Prop({ mutable: true }) public activeIdentifier?: string | undefined;
+  @Prop() public activeIdentifier?: string | undefined;
 
   /** Adapts the flyout-navigation color depending on the theme. */
   @Prop() public theme?: Theme = 'light';
@@ -56,13 +56,13 @@ export class FlyoutNavigation {
   /** Emitted when activeIdentifier is changed. */
   @Event({ bubbles: false }) public update?: EventEmitter<FlyoutNavigationUpdateEvent>;
 
+  @State() private flyoutNavigationItemElements: HTMLPFlyoutNavigationItemElement[] = [];
+
   private dialog: HTMLDialogElement;
-  private flyoutNavigationItemElements: HTMLPFlyoutNavigationItemElement[] = [];
 
   @Watch('open')
   public openChangeHandler(isOpen: boolean): void {
     setScrollLock(isOpen);
-    this.setDialogVisibility(isOpen);
   }
 
   public componentWillLoad(): void {
@@ -71,7 +71,6 @@ export class FlyoutNavigation {
     this.host.shadowRoot.addEventListener(INTERNAL_UPDATE_EVENT_NAME, (e: CustomEvent<FlyoutNavigationUpdateEvent>) => {
       e.stopPropagation(); // prevents internal event from bubbling further
       const activeIdentifier = e.detail.activeIdentifier;
-      this.activeIdentifier = activeIdentifier;
       this.update.emit({ activeIdentifier });
     });
   }
@@ -90,6 +89,11 @@ export class FlyoutNavigation {
 
   public disconnectedCallback(): void {
     setScrollLock(false);
+  }
+
+  public componentDidRender(): void {
+    // showModal needs to be called after render cycle to prepare visibility states of dialog in order to focus the dismiss button correctly
+    this.setDialogVisibility(this.open);
   }
 
   public render(): JSX.Element {
@@ -153,9 +157,11 @@ export class FlyoutNavigation {
   }
 
   private setDialogVisibility(isOpen: boolean): void {
-    if (isOpen === true) {
+    // TODO: SupportsNativeDialog check
+    // Only call showModal/close on dialog when state changes
+    if (isOpen === true && !this.dialog.open) {
       this.dialog.showModal();
-    } else {
+    } else if (isOpen === false && this.dialog.open) {
       this.dialog.close();
     }
   }